Solve circular dependacy and load edit/create forms bugs
