+++ src/button_action.js
@@ -1,18 +1,20 @@
-import getModal from './modal';
 import { getTodo } from './data';
 
 const addTodoBtnAction = (projectId) => {
-  const modalContainer = document.getElementById('modal-box');
-  modalContainer.innerHTML = '';
-  modalContainer.appendChild(getModal('todo'));
+  document.forms['new-todo'].querySelector('#todo-title-input').value = '';
+  document.forms['new-todo'].querySelector('#todo-desc-input').value = '';
+  document.forms['new-todo'].querySelector('#todo-duedate-input').value = '';
+  document.forms['new-todo'].querySelector('#todo-low-priority').checked = false;
+  document.forms['new-todo'].querySelector('#todo-medium-priority').checked = true;
+  document.forms['new-todo'].querySelector('#todo-high-priority').checked = false;
+  document.forms['new-todo'].querySelector('#todo-notes-input').value = '';
   document.forms['new-todo'].querySelector('#project-input').value = projectId;
-  document.getElementById('form-box').classList.toggle('show-form');
+  document.forms['new-todo'].querySelector('#todo-input').value = -1;
+  document.forms['new-todo'].querySelector('input[type=submit]').value = 'Create New Todo';
+  document.getElementById('todo-form-box').classList.toggle('show-form');
 };
 
 const editTodoBtnAction = (todoId, projectId) => {
-  const modalContainer = document.getElementById('modal-box');
-  modalContainer.innerHTML = '';
-  modalContainer.appendChild(getModal('todo'));
   const todo = getTodo(todoId, projectId);
   const {
     title,
@@ -29,8 +31,8 @@ const editTodoBtnAction = (todoId, projectId) => {
   document.forms['new-todo'].querySelector('#todo-notes-input').value = notes;
   document.forms['new-todo'].querySelector('#project-input').value = projectId;
   document.forms['new-todo'].querySelector('#todo-input').value = todoId;
-  document.forms['new-todo'].querySelector('input[type=submit]').value = 'Update Todo Details';
-  document.getElementById('form-box').classList.toggle('show-form');
+  document.forms['new-todo'].querySelector('input[type=submit]').value = 'Update Todo';
+  document.getElementById('todo-form-box').classList.toggle('show-form');
 };
 
 export { addTodoBtnAction, editTodoBtnAction };
